+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next release
+## [1.9.4](https://github.com/sdss/sdsstools/compare/1.9.3...1.9.4) - 2025-10-23
 
 - Fix issues casting YAML data to `return_class` in `read_yaml_file`.